Extract test stub view to its own module
