+++ prlint/github/tests/test_payload_factories/test_payload_request_factory.py
@@ -1,21 +1,7 @@
 import unittest
 
-from rest_framework.decorators import api_view
-from rest_framework.response import Response
-from rest_framework.views import APIView
-
 from ..payload_factories import PayloadRequestFactory
-
-
-@api_view(['POST'])
-def view(request):
-    """
-    Testing stub view to return Request's data and GitHub event header.
-    """
-    return Response({
-        'header_github_event': request.META.get('HTTP_X_GITHUB_EVENT', ''),
-        'request_data': request.data,
-    })
+from .view_stubs import view
 
 
 class TestPayloadRequestFactory(unittest.TestCase):
