+++ prlint/github/tests/views/test_post.py
@@ -70,9 +70,6 @@ class TestPostRegistered(TestPost):
 
         self.assertEqual(result.status_code, status.HTTP_403_FORBIDDEN)
         self.assertEqual(list(result.data), ['detail'])
-        {
-            "detail": "The 'issues' event is not accepted by this webhook. Please reconfigure.",
-        }
         self.assertIn('\'commit\' event', result.data['detail'])
         self.assertIn('Please reconfigure.', result.data['detail'])
